Refactor: LadderFactory, LineFactory 분리
