+++ src/main/java/nextstep/ladder/domain/Ladder.java
@@ -2,28 +2,23 @@ package nextstep.ladder.domain;
 
 import nextstep.ladder.validator.LadderValidator;
 
-import java.util.*;
-import java.util.stream.Collectors;
-import java.util.stream.Stream;
+import java.util.Collections;
+import java.util.LinkedHashMap;
+import java.util.List;
+import java.util.Map;
 
 public class Ladder {
 
     public static final int MIN_LADDER_HEIGHT = 1;
     private final List<Line> lines;
 
+    public static Ladder of(List<Line> lines) {
+        return new Ladder(lines);
+    }
+
     private Ladder(List<Line> lines) {
         LadderValidator.validateLadderMaxHeight(lines.size());
         this.lines = lines;
-    }
-
-    public static Ladder of(int numberOfParticipant, int height) {
-        return new Ladder(createLines(numberOfParticipant, height));
-    }
-
-    private static List<Line> createLines(int pointSize, int height) {
-        return Stream.generate(() -> LineFactory.createLine(pointSize, height))
-                .limit(height)
-                .collect(Collectors.toList());
     }
 
     public List<Line> getLines() {
